+++ src/main/kotlin/io/kadena/pact/PactIcons.kt
@@ -1,0 +1,10 @@
+package io.kadena.pact
+
+import com.intellij.openapi.util.IconLoader
+import javax.swing.Icon
+
+public object PactIcons {
+
+    @JvmField
+    public val FILE: Icon = IconLoader.getIcon("/icons/file.svg", javaClass)
+}