+++ src/main/kotlin/io/kadena/pact/PactLanguage.kt
@@ -1,0 +1,11 @@
+package io.kadena.pact
+
+import com.intellij.lang.Language
+
+public class PactLanguage : Language("Pact") {
+
+    companion object {
+        @JvmStatic
+        public val INSTANCE = PactLanguage()
+    }
+}