+++ src/main/kotlin/io/kadena/pact/ide/settings/AppSettingsState.kt
@@ -1,0 +1,54 @@
+package io.kadena.pact.ide.settings
+
+import com.intellij.openapi.application.ApplicationManager
+import com.intellij.openapi.components.PersistentStateComponent
+import com.intellij.openapi.components.State
+import com.intellij.openapi.components.Storage
+import com.intellij.util.xmlb.XmlSerializerUtil
+import org.jetbrains.annotations.NotNull
+import org.jetbrains.annotations.Nullable
+
+
+/**
+ * Supports storing the application settings in a persistent way.
+ * The [State] and [Storage] annotations define the name of the data and the file name where
+ * these persistent application settings are stored.
+ */
+@State(
+    name = "io.kadena.pact.ide.settings.AppSettingsState",
+    storages = [Storage("PactPlugin.xml")]
+)
+class AppSettingsState : PersistentStateComponent<AppSettingsState> {
+    var pactPath: String = ""
+    var pactLanguageServerPath: String = ""
+
+    private val listeners = mutableSetOf<AppSettingsStateListener>()
+
+    @Nullable
+    override fun getState(): AppSettingsState {
+        return this
+    }
+
+    override fun loadState(@NotNull state: AppSettingsState) {
+        XmlSerializerUtil.copyBean(state, this)
+    }
+
+    companion object {
+        val instance: AppSettingsState
+            get() = ApplicationManager.getApplication().getService(AppSettingsState::class.java)
+    }
+
+    fun addAppSettingsStateListener(listener: AppSettingsStateListener) {
+        listeners.add(listener)
+    }
+
+    fun notifyAppSettingsStateChanged(state: AppSettingsState) {
+        for (listener in listeners) {
+            listener.onAppSettingsStateChanged(state)
+        }
+    }
+
+    interface AppSettingsStateListener {
+        fun onAppSettingsStateChanged(state: AppSettingsState)
+    }
+}