+++ CHANGELOG.md
@@ -4,4 +4,4 @@
 
 ## [Unreleased]
 ### Added
-- Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)
+- Initial support for the [Pact Language Server (LSP)](https://github.com/kadena-io/pact-lsp)