+++ CHANGELOG.md
@@ -3,5 +3,12 @@
 # pact-intellij Changelog
 
 ## [Unreleased]
+
+## [0.1.0] - 2024-01-22
+
 ### Added
-- Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)
+
+- Initial support for the [Pact Language Server (LSP)](https://github.com/kadena-io/pact-lsp)
+
+[Unreleased]: https://github.com/lukeribchester/pact-intellij/compare/v0.1.0...HEAD
+[0.1.0]: https://github.com/lukeribchester/pact-intellij/commits/v0.1.0